+++ src/avp/types/assigned_tunnel_id.rs
@@ -8,10 +8,10 @@ pub struct AssignedTunnelId {
 
 impl AssignedTunnelId {
     const ATTRIBUTE_TYPE: u16 = 9;
-    const LENGTH: u16 = 2;
+    const LENGTH: usize = 2;
 
     pub fn try_read<'a, 'b>(reader: &'b mut impl Reader<'a>) -> ResultStr<Self> {
-        if reader.len() < Self::LENGTH as usize {
+        if reader.len() < Self::LENGTH {
             return Err("Incomplete AssignedTunnelId AVP encountered");
         }
 
@@ -27,7 +27,7 @@ impl From<u16> for AssignedTunnelId {
 }
 
 impl QueryableAVP for AssignedTunnelId {
-    fn get_length(&self) -> u16 {
+    fn get_length(&self) -> usize {
         Self::LENGTH
     }
 }
