+++ src/avp/types/initial_received_lcp_conf_req.rs
@@ -21,8 +21,8 @@ impl InitialReceivedLcpConfReq {
 }
 
 impl QueryableAVP for InitialReceivedLcpConfReq {
-    fn get_length(&self) -> u16 {
-        self.value.len() as u16
+    fn get_length(&self) -> usize {
+        self.value.len()
     }
 }
