+++ src/avp/types/protocol_version.rs
@@ -8,10 +8,10 @@ pub struct ProtocolVersion {
 }
 
 impl ProtocolVersion {
-    const LENGTH: u16 = 2;
+    const LENGTH: usize = 2;
 
     pub fn try_read<'a, 'b>(reader: &'b mut impl Reader<'a>) -> ResultStr<Self> {
-        if reader.len() < Self::LENGTH as usize {
+        if reader.len() < Self::LENGTH {
             return Err("Incomplete ProtocolVersion AVP encountered");
         }
 
@@ -22,7 +22,7 @@ impl ProtocolVersion {
 }
 
 impl QueryableAVP for ProtocolVersion {
-    fn get_length(&self) -> u16 {
+    fn get_length(&self) -> usize {
         Self::LENGTH
     }
 }
