+++ src/avp/types/receive_window_size.rs
@@ -7,10 +7,10 @@ pub struct ReceiveWindowSize {
 }
 
 impl ReceiveWindowSize {
-    const LENGTH: u16 = 2;
+    const LENGTH: usize = 2;
 
     pub fn try_read<'a, 'b>(reader: &'b mut impl Reader<'a>) -> ResultStr<Self> {
-        if reader.len() < Self::LENGTH as usize {
+        if reader.len() < Self::LENGTH {
             return Err("Incomplete ReceiveWindowSize AVP encountered");
         }
 
@@ -26,7 +26,7 @@ impl From<u16> for ReceiveWindowSize {
 }
 
 impl QueryableAVP for ReceiveWindowSize {
-    fn get_length(&self) -> u16 {
+    fn get_length(&self) -> usize {
         Self::LENGTH
     }
 }
